Add a separate file for signing room management
